Fix max time issues that come with ...
enabling or disableing bomb casings with at_bash's mod selector.

(The bug is that if Quad decker was disabled, then you enabled it,
instead of max time going from 20:00 to 92:00, max time goes from 20:00
to 56:00.   Ditto in reverse,  if Quad decker was enabled on start then
disabled later,  Max time would go from 92:00 to 56:00 instead of 92:00
to 20:00.)

Ditto with Double decker,  (time from 20:00 to 32:00 instead of 20:00 to
44:00),  or from double to Quad decker,  (44:00 to 68:00 instead of
44:00 to 92:00).